+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fix usage documentation of `ItemSearch`
 - Fix fields argument to CLI  ([#797](https://github.com/stac-utils/pystac-client/pull/797))
+- Clarify recursive behaviour of the `get_items` method in the method docstring ([#800](https://github.com/stac-utils/pystac-client/pull/800))
 
 ## [v0.8.6] - 2025-02-11